refactor: rename profile components for clarity
- Renamed Profile.svelte to EditProfile.svelte (edit functionality)
- Renamed ParticipantProfile.svelte to Profile.svelte (public view)
- Updated route definitions: /profile -> EditProfile, /participant/:address -> Profile
- Updated ValidatorWaitlist to redirect to public profile after joining
- Updated documentation with new component names
